remove handler::request module.
