+++ web/src/handler/types/mod.rs
@@ -3,7 +3,6 @@ pub mod extension;
 pub mod header;
 pub mod html;
 pub mod path;
-pub mod request;
 pub mod state;
 pub mod string;
 pub mod uri;
